[staccmr-21] moved files to cmr directory
