+++ search/lib/cmr/cmr_converter.js
@@ -1,5 +1,5 @@
 const _ = require('lodash');
-const { wfs, generateAppUrl, extractParam, generateSelfUrl } = require('./util');
+const { wfs, generateAppUrl, extractParam, generateSelfUrl } = require('../util');
 const cmr = require('./cmr');
 
 const WHOLE_WORLD_BBOX = [-180, 90, 180, -90];
